pass endpointUri to resource, fix paginationurl
getPaginationUrl now requires a String endpoint to build the url from. Also deprecated the gateway url since it was redundant. Fixed some of the unit tests and added a return value for getMatchedUris although that @Override will probably have to be changed to something more dynamic in the future if we want to test against it. For now, it just needs to return something