added .json() that was missing from some requests
also removed skip regarding apigee cache issue. This has been fixed. However I did add a skip because, given the version of dw we're running as of now (1.0.2), a request using a not allowed methods returns a 405 to apigee but without any allowed methods headers, so apigee returns a 502-bad-gateway. When DW releases 1.10, it will include a fix for the missing headers issue