use ElasticSearch java library
CO-927
CO-721

Instead of manually making HTTP requests to the REST API, which is
apparently not very stable, use the java library to make requests.

This commit only changes the getById and getServiceById methods, since
they are the least complicated and easiest to change. Other methods will
be changed in future commits.