Fix sort order of location results
CO-813

When searching for locations, we want the results to be in an order
which puts the best match first; however, they were being returned
unordered.

The reason is that we were applying the search terms with a bool filter
query, which, according to the documentation, is the same as a 'must'
query except [1]

> However unlike 'must' the score of the query will be ignored

So even though elasticsearch was capable of calculating how well the
search terms matched, this wasn't contributing to the overall score
because we were telling it not to.

This commit fixes that by using a 'must' query instead of a 'filter'
query. It also explicitly sets the fallback sort order to '_score', just
in case.

[1]: https://www.elastic.co/guide/en/elasticsearch/reference/2.4/query-dsl-bool-query.html