Use AND instead of OR for search queries
CO-813

A search for "milam hall" would return all locations which matched
"milam" OR "hall", which is elasticsearch's default. Change this so that
the same search matches "milam" AND "hall".

Note: the way the AND operator interacts with a multi_match query is
that all the search terms must be found in *one* of the search fields.
This seems okay for now.

https://www.elastic.co/guide/en/elasticsearch/reference/2.4/query-dsl-multi-match-query.html#operator-min