only get parking locations
if the random number generated was 0, it would also return locations with null counts for moto/ev/ada spaces. This would cause the test to throw an error when comparing a number to null. All parking locations should have a number for the ev/ada/moto fields, so getting only parking data tests that these fields are never null.